🚚 refactor: move shared functions to common utilities folder
